dom/inputmethod/mochitest/test_inputMethodManager.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=955965
-->
<head>
  <title>Test for InputMethodManager</title>
  <script type="application/javascript;version=1.7" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.7" src="inputmethod_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=955965">Mozilla Bug 955965</a>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.7">

let gIm;

inputmethod_setup(function() {
  runTest();
});

function runTest() {
  gIm = navigator.mozInputMethod;
  gIm.oninputcontextchange = function() {
    ok(true, 'inputcontextchange event was fired.');
    gIm.oninputcontextchange = null;

    if (!gIm.inputcontext) {
      ok(false, 'Should have a non-null inputcontext.');
      inputmethod_cleanup();
      return;
    }

    test_hide();
  };

  // Set current page as an input method.
  SpecialPowers.wrap(gIm).setActive(true);

  let iframe = document.createElement('iframe');
  iframe.src = 'file_test_app.html';
  iframe.setAttribute('mozbrowser', true);
  document.body.appendChild(iframe);
}

function test_hide() {
  gIm.oninputcontextchange = function() {
    is(gIm.inputcontext, null, 'Should have a null inputcontext.');
    inputmethod_cleanup();
  }
  gIm.mgmt.hide();
}

</script>
</pre>
</body>
</html>
